Revert "A bad commit!"
This reverts commit 700af3dc5156c2a1c1d61532f8f3b22329630443.

I was showing a student an example of the commit process, and then thought it might be kind of smart to demonstrate "reverting" a commit, that is undoing it so that its changes are no longer in the files, but it still appears as an event in the history of the repository.

Essentially this is like a highly visible "undo".